Extract some functions from godot_api.rs into own files
